Use Int progress callback; simplify CopyFiles UI
Change onProgressUpdate to use Int in ExecuteCopyFiles and update the file handler to pass an Int count. Adjust CopyFilesView to initialize max to 0, replace the custom ProgressCount usage with a simple inline ProgressView/Text, and hide action buttons while a copy is in progress instead of merely disabling them. Wire the new Int callback through a @MainActor Task and convert to Double for the view's progress state via a small onProgressUpdate helper. Minor whitespace cleanup in ProgressCount.swift.

+++ RawCull/Model/ParametersRsync/ExecuteCopyFiles.swift
@@ -42,7 +42,7 @@ final class ExecuteCopyFiles {
     private var destAccessedURL: URL?
 
     // Callbacks
-    var onProgressUpdate: ((Double) -> Void)?
+    var onProgressUpdate: ((Int) -> Void)?
     var onCompletion: ((CopyDataResult) -> Void)?
 
     func startcopyfiles(
@@ -153,7 +153,7 @@ final class ExecuteCopyFiles {
             fileHandler: { [weak self] count in
                 Task { @MainActor in
                     if let update = self?.onProgressUpdate {
-                        update(Double(count))
+                        update(count)
                     }
                 }
             },
